backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-socketio==5.11.4
python-multipart==0.0.12
websockets==13.1
twilio==9.3.3
openai==1.51.0
gspread==6.1.4
google-auth==2.35.0
numpy==2.1.2
scipy==1.14.1
structlog==24.4.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
qdrant-client>=1.9.0,<2.0.0
PyJWT==2.9.0
audioop-lts; python_version >= "3.13"
